Crate yolofd

A tiny, synchronous multipart/form-data writer.

This is mostly untested and was not written with quality in mind. It works for the one use case that I need it for :)

FormData::new() is the entry point for this library.

Structs

Field

FieldBuilder
Builder for Field values.

FormData
A multipart/form-data format writer.

Functions

generate_boundary
Generate a random string that can be used as a multipart boundary.